requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==6.0
django-ckeditor-5==0.2.18
django-crispy-forms==2.5
django-js-asset==3.1.2
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
